Added OutlookFeatureHasBeenDisabledPatch.xml
Added OutlookFeatureHasBeenDisabledPatch.xml which suppresses the pop up window in outlook stating "this feature has been disabled by your system administrator" and the subsequent "need password" button on the bottom left of the windows outlook application. This issue is present when there is an On-Premises exchange environment and the outlook client is attempting to reach out to cloud hosted exchange and the use online content options have been set to disabled.